fix(ci): restore desktop bindings and satisfy repolint
Problem
CI lint failed on comment, file-size, and complexity overruns. Desktop
jobs failed TypeScript because Wails still exports deprecated
SetAgentPreset*/SetTokenMode* and DeliveryWorktree aliases that
AppBindings no longer listed.

Root cause
The standard-execution change kept the one-version Go compatibility
methods bound, but the frontend drift check requires every generated
Wails method on AppBindings. The same change also grew several
already-baselined files past comment and size budgets.

Fix
Bind the deprecated Wails methods as no-op aliases in AppBindings and
the browser mock. Extract ExecutionPolicyBlock and boot test helpers,
split Derive helpers, and compress comments and tests so repolint stays
inside existing budgets.

Verification
- go run ./tools/repolint
- go test ./internal/taskpolicy ./internal/acp ./internal/boot ./internal/control ./internal/evidence ./internal/cli ./internal/capability ./internal/agent
- cd desktop && go test . -run 'TestLoadTabSessionProfile|TestSetTokenMode|TestSetAgentPreset|TestDeliveryWorktree|TestCreateDelivery|TestIsolated'
- desktop/frontend tsc --noEmit
- isolated-worktree and composer-profile frontend tests
