+++ internal/control/turn_orchestrator_test.go
@@ -70,11 +70,8 @@ func TestTurnOrchestratorAttachesTrustedPlannerMetadata(t *testing.T) {
 	if runner.meta.UserText != raw {
 		t.Fatalf("planner metadata user text = %q, want pristine %q", runner.meta.UserText, raw)
 	}
-	if runner.meta.ExplicitPlanMode || !runner.meta.GoalActive || !runner.meta.ClosedLoop {
+	if runner.meta.ExplicitPlanMode || !runner.meta.GoalActive || !runner.meta.ClosedLoop || !runner.meta.PolicySet || !runner.meta.Policy.ClosedLoop() {
 		t.Fatalf("planner metadata missing trusted host state: %+v", runner.meta)
-	}
-	if !runner.meta.PolicySet || !runner.meta.Policy.ClosedLoop() {
-		t.Fatalf("planner metadata should carry a closed-loop TaskPolicy for an active Goal: %+v", runner.meta)
 	}
 	if !runner.meta.HasConversationContext {
 		t.Fatalf("planner metadata lost executor conversation ownership: %+v", runner.meta)
