+++ internal/evidence/evidence.go
@@ -1418,16 +1418,12 @@ func FromContext(ctx context.Context) (*Ledger, bool) {
 	return ledger, ok && ledger != nil
 }
 
-// WithClosedLoopExecution marks tool execution as subject to the closed-loop
-// final-readiness contract. Tools use this only for stricter evidence
-// validation; it is ephemeral host state and is never serialized into sessions
-// or prompts.
+// WithClosedLoopExecution marks a tool call for closed-loop evidence checks.
 func WithClosedLoopExecution(ctx context.Context) context.Context {
 	return context.WithValue(ctx, closedLoopKey{}, true)
 }
 
-// ClosedLoopExecutionFromContext reports whether the current tool call must
-// produce evidence that the closed-loop final-readiness gate can accept.
+// ClosedLoopExecutionFromContext reports whether closed-loop evidence is required.
 func ClosedLoopExecutionFromContext(ctx context.Context) bool {
 	enabled, _ := ctx.Value(closedLoopKey{}).(bool)
 	return enabled
@@ -1576,9 +1572,7 @@ func ToolCallPaths(args json.RawMessage) []string {
 	return out
 }
 
-// ToolCallRequiresAcceptanceCriteria reports whether a call begins execution
-// work that needs an acceptance contract. Mutations always qualify; verification
-// commands also qualify even though they are intentionally not mutations.
+// ToolCallRequiresAcceptanceCriteria reports mutations and verification commands.
 func ToolCallRequiresAcceptanceCriteria(toolName string, args json.RawMessage, readOnly bool) bool {
 	if ToolCallMutates(toolName, args, readOnly) {
 		return true
@@ -1860,10 +1854,7 @@ func bashStaticArgv(command string) ([]string, bool) {
 	return fields, malformed == "" && len(fields) > 0
 }
 
-// IsVerificationCommand reports whether command is a host-recognized
-// verification command for evidence closure. Keep complete_step and the
-// final-readiness gate on this single classifier so a sign-off cannot claim a
-// command that the final gate will immediately reject.
+// IsVerificationCommand reports whether command is a recognized verifier.
 func IsVerificationCommand(command string) bool {
 	return bashCommandIsVerification(command)
 }
